Add .gitignore and differential_base_alu model
Added a .gitignore file in the .idea directory to ignore files that are not needed in the repository, such as local IDE settings and data sources. This will make the repo cleaner and easier to manage.
Additionally, a new 3D model for the differential base (differential_base_alu.dae) is added to the models directory. The model will be used in the visualization of the robot base.