Refactor the gz-sim-diff-drive-system plugin position in urdf file
The gz-sim-diff-drive-system plugin block is relocated earlier up in the minipock.urdf.xacro file. This adjustment is made for enhanced readability and maintainability. It organizes the sections very clearly by calling the gz-sim-diff-drive-system plugin before the PosePublisher and the JointStatePublisher plugins. This further assists developers by implying the order of plugin usage or requirement.