Update minipock_world simulation parameters
A new name for the world was set from 'industrial-warehouse' to 'minipock_world' and the pose of the ground model in the world was also adjusted from -0.090092 to -0.2 in order to better cater to the MiniPock bot.

These changes will improve the simulation environment for the MiniPock bot, providing a more realistic testing and development platform.