Refactor spawn.launch.py, add map_server and bridge functions
The spawn.launch.py was refactored by reordering imports and organizing the code. This was done to make the code clearer and more structured for reuse and maintenance. New functions like map_server and bridge were added. The map_server method will allow to easily setup and manage the map server for simulations, ultimately improving the mapping functionality in the simulation. The bridge function brings in the benefits of using ROS2 to ROS1 bridge, so any ROS1 packages can communicate with ROS2 ones, enhancing intercompatibility. Additionally, reorganizations in the spawn function are performed to accommodate these updates.