Swap Poetry packaging with traditional setuptools
The use of Python packaging tool Poetry has been abandoned in favor of a more traditional approach using setuptools. This was implemented through the deletion of both poetry.lock and pyproject.toml files, which was specific to Poetry-based projects. Instead, we have introduced setup.py for package configuration and dependency management and setup.cfg for script locations.