Add visual configuration for cartographer mapping
Introduced visual configuration for the minipock_cartographer RVIZ display setup. This includes configurations, launch scripts and the relevant settings for the grid cells, map frames, range finders and different topics. A new node for the Cartographer is launched along with occupancy grid node and the RVIZ node depending on conditions.