Update arguments in spawn_args function call
This commit involves modifying the spawn_args function call in the Node attribute of our 'ros_gz_sim' package. The function call's position argument was removed for better compatibility and reduced complexity. This streamlines the process of spawning launch files in the 'minipock_description' package and makes the system less prone to position-related errors.